+++ src/xrParticles/psystem.h
@@ -208,7 +208,7 @@ public:
 
     // action
     virtual ParticleAction* CreateAction(PActionEnum type) = 0;
-    virtual u32 LoadActions(int alist_id, IReader& R) = 0;
+    virtual size_t LoadActions(int alist_id, IReader& R) = 0;
     virtual void SaveActions(int alist_id, IWriter& W) = 0;
 };
